requirements.txt
fastapi
uvicorn
loguru
PyYAML
SQLAlchemy
aiosqlite
APScheduler
pydantic
jinja2
cryptography
python-dotenv
httpx